Create AtomMatrix_14
This is an update to the base code that brings the scanning in-line with lukeswitz's 1.40b3 masterpiece.

I've tuned the LED code to be way more efficient and better reflect the activity that is happening. This appears to have helped with speed tremendously - I am not seeing roughly equivalent finds between Luke's code on a regular AtomGPS and mine on the Matrix.